Remove temp app.py with current (not working) one, and added login route inside auth blueprint
- Removed temporary app.py and replaced it with old "appCurrent.py" file
- Added login route(s) to auth.py
- Commented blueprint imports added to app.py
